docker/mcp/requirements.mcp.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain>=0.1.0
langchain-mistralai
pyodbc
qdrant-client==1.15.1
msal
mcp
sqlalchemy
aioodbc
pymssql